[refactor] Fixed test after refactor
