+++ test/test_orm.py
@@ -6,20 +6,19 @@ import pycram.orm.action_designator
 import pycram.orm.base
 import pycram.orm.motion_designator
 import pycram.orm.object_designator
-import pycram.orm.task
-import pycram.task
-import pycram.task
+import pycram.orm.tasktree
+import pycram.tasktree
 from bullet_world_testcase import BulletWorldTestCase
 from pycram.world_concepts.world_object import Object
 from pycram.designators import action_designator, object_designator, motion_designator
-from pycram.designators.actions.actions import ParkArmsActionPerformable, MoveTorsoActionPerformable, \
+from pycram.designators.performables.actions import ParkArmsActionPerformable, MoveTorsoActionPerformable, \
     SetGripperActionPerformable, PickUpActionPerformable, NavigateActionPerformable, TransportActionPerformable, \
     OpenActionPerformable, CloseActionPerformable, DetectActionPerformable, LookAtActionPerformable
 from pycram.designators.object_designator import BelieveObject
 from pycram.datastructures.enums import ObjectType
 from pycram.datastructures.pose import Pose
 from pycram.process_module import simulated_robot
-from pycram.task import with_tree
+from pycram.tasktree import with_tree
 from pycram.orm.views import PickUpWithContextView
 
 
@@ -39,7 +38,7 @@ class DatabaseTestCaseMixin(BulletWorldTestCase):
 
     def tearDown(self):
         super().tearDown()
-        pycram.task.reset_tree()
+        pycram.tasktree.reset_tree()
         pycram.orm.base.ProcessMetaData.reset()
         pycram.orm.base.Base.metadata.drop_all(self.engine)
         self.session.close()
@@ -84,10 +83,10 @@ class ORMTaskTreeTestCase(DatabaseTestCaseMixin):
         """Test if the objects in the database is equal with the objects that got serialized."""
         self.plan()
         pycram.orm.base.ProcessMetaData().description = "Unittest"
-        pycram.task.task_tree.root.insert(self.session, )
+        pycram.tasktree.task_tree.root.insert(self.session, )
 
-        node_results = self.session.scalars(select(pycram.orm.task.TaskTreeNode)).all()
-        self.assertEqual(len(node_results), len(pycram.task.task_tree.root))
+        node_results = self.session.scalars(select(pycram.orm.tasktree.TaskTreeNode)).all()
+        self.assertEqual(len(node_results), len(pycram.tasktree.task_tree.root))
 
         position_results = self.session.scalars(select(pycram.orm.base.Position)).all()
         self.assertEqual(14, len(position_results))
@@ -107,22 +106,22 @@ class ORMTaskTreeTestCase(DatabaseTestCaseMixin):
     def test_metadata_existence(self):
         pycram.orm.base.ProcessMetaData().description = "metadata_existence_test"
         self.plan()
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(pycram.orm.base.Pose)).all()
         self.assertTrue(all([r.process_metadata is not None for r in result]))
 
     def test_task_tree_node_parents(self):
         self.plan()
         pycram.orm.base.ProcessMetaData().description = "task_tree_node_parents_test"
-        pycram.task.task_tree.root.insert(self.session)
-        result = self.session.scalars(select(pycram.orm.task.TaskTreeNode)).all()
+        pycram.tasktree.task_tree.root.insert(self.session)
+        result = self.session.scalars(select(pycram.orm.tasktree.TaskTreeNode)).all()
         self.assertTrue([result[i].parent == result[result[i].parent_id - 1] for i in range(len(result))
                          if result[i].parent_id is not None])
 
     def test_meta_data(self):
         self.plan()
         pycram.orm.base.ProcessMetaData().description = "Unittest"
-        pycram.task.task_tree.root.insert(self.session, )
+        pycram.tasktree.task_tree.root.insert(self.session, )
         metadata_results = self.session.scalars(select(pycram.orm.base.ProcessMetaData)).all()
         self.assertEqual(1, len(metadata_results))
 
@@ -138,7 +137,7 @@ class ORMTaskTreeTestCase(DatabaseTestCaseMixin):
     def test_meta_data_alternation(self):
         self.plan()
         pycram.orm.base.ProcessMetaData().description = "meta_data_alternation_test"
-        pycram.task.task_tree.root.insert(self.session, )
+        pycram.tasktree.task_tree.root.insert(self.session, )
         metadata_result = self.session.scalars(select(pycram.orm.base.ProcessMetaData)).first()
         self.assertEqual(metadata_result.description, "meta_data_alternation_test")
 
@@ -158,14 +157,14 @@ class MixinTestCase(DatabaseTestCaseMixin):
     def test_pose(self):
         self.plan()
         pycram.orm.base.ProcessMetaData().description = "pose_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(pycram.orm.base.Pose)).all()
         self.assertTrue(all([r.position is not None and r.orientation is not None for r in result]))
 
     def test_pose_mixin(self):
         self.plan()
         pycram.orm.base.ProcessMetaData().description = "pose_mixin_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(pycram.orm.base.RobotState)).all()
         self.assertTrue(all([r.pose is not None and r.pose_id == r.pose.id for r in result]))
 
@@ -183,7 +182,7 @@ class ORMObjectDesignatorTestCase(DatabaseTestCaseMixin):
             PickUpActionPerformable(object_description.resolve(), "left", "front").perform()
             description.resolve().perform()
         pycram.orm.base.ProcessMetaData().description = "Unittest"
-        tt = pycram.task.task_tree
+        tt = pycram.tasktree.task_tree
         tt.insert(self.session)
         action_results = self.session.scalars(select(pycram.orm.action_designator.Action)).all()
         motion_results = self.session.scalars(select(pycram.orm.motion_designator.Motion)).all()
@@ -197,8 +196,8 @@ class ORMActionDesignatorTestCase(DatabaseTestCaseMixin):
         with simulated_robot:
             action.perform()
         pycram.orm.base.ProcessMetaData().description = "code_designator_type_test"
-        pycram.task.task_tree.root.insert(self.session)
-        result = self.session.scalars(select(pycram.orm.task.TaskTreeNode).where(pycram.orm.task.TaskTreeNode.
+        pycram.tasktree.task_tree.root.insert(self.session)
+        result = self.session.scalars(select(pycram.orm.tasktree.TaskTreeNode).where(pycram.orm.tasktree.TaskTreeNode.
                                                                                  action_id.isnot(None))).all()
         self.assertEqual(result[0].action.dtype, action_designator.NavigateAction.__name__)
         self.assertEqual(result[1].action.dtype, motion_designator.MoveMotion.__name__)
@@ -208,7 +207,7 @@ class ORMActionDesignatorTestCase(DatabaseTestCaseMixin):
         with simulated_robot:
             action.perform()
         pycram.orm.base.ProcessMetaData().description = "parkArmsAction_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(pycram.orm.action_designator.ParkArmsAction)).all()
         self.assertTrue(all([result[i + 1].dtype is not pycram.orm.action_designator.Action.dtype
                              if result[i].dtype is pycram.orm.action_designator.ParkArmsAction.dtype else None
@@ -221,7 +220,7 @@ class ORMActionDesignatorTestCase(DatabaseTestCaseMixin):
         with simulated_robot:
             action.perform()
         pycram.orm.base.ProcessMetaData().description = "transportAction_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(pycram.orm.action_designator.TransportAction)).all()
         milk_object = self.session.scalars(select(pycram.orm.object_designator.Object)).first()
         self.assertEqual(milk_object.pose, result[0].object.pose)
@@ -235,7 +234,7 @@ class ORMActionDesignatorTestCase(DatabaseTestCaseMixin):
             LookAtActionPerformable(object_description.resolve().pose).perform()
             action.perform()
         pycram.orm.base.ProcessMetaData().description = "detectAction_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(pycram.orm.action_designator.DetectAction)).all()
         self.assertEqual(result[0].object.name, "milk")
 
@@ -244,7 +243,7 @@ class ORMActionDesignatorTestCase(DatabaseTestCaseMixin):
         with simulated_robot:
             action.perform()
         pycram.orm.base.ProcessMetaData().description = "setGripperAction_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(pycram.orm.action_designator.SetGripperAction)).all()
         self.assertEqual(result[0].gripper, "left")
         self.assertEqual(result[0].motion, "open")
@@ -264,7 +263,7 @@ class ORMActionDesignatorTestCase(DatabaseTestCaseMixin):
             CloseActionPerformable(handle_desig, arm="left").perform()
 
         pycram.orm.base.ProcessMetaData().description = "open_and_closeAction_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         open_result = self.session.scalars(select(pycram.orm.action_designator.OpenAction)).all()
         close_result = self.session.scalars(select(pycram.orm.action_designator.CloseAction)).all()
         self.assertTrue(open_result is not None)
@@ -277,7 +276,7 @@ class ORMActionDesignatorTestCase(DatabaseTestCaseMixin):
 class ViewsSchemaTest(DatabaseTestCaseMixin):
     def test_view_creation(self):
         pycram.orm.base.ProcessMetaData().description = "view_creation_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         view = PickUpWithContextView
         self.assertEqual(len(view.__table__.columns), 12)
         self.assertEqual(view.__table__.name, "PickUpWithContextView")
@@ -304,7 +303,7 @@ class ViewsSchemaTest(DatabaseTestCaseMixin):
             PickUpActionPerformable(object_description.resolve(), "left", "front").perform()
             description.resolve().perform()
         pycram.orm.base.ProcessMetaData().description = "pickUpWithContextView_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(PickUpWithContextView)).first()
         self.assertEqual(result.arm, "left")
         self.assertEqual(result.grasp, "front")
@@ -324,7 +323,7 @@ class ViewsSchemaTest(DatabaseTestCaseMixin):
             PickUpActionPerformable(object_description.resolve(), "left", "front").perform()
             description.resolve().perform()
         pycram.orm.base.ProcessMetaData().description = "pickUpWithContextView_conditions_test"
-        pycram.task.task_tree.root.insert(self.session)
+        pycram.tasktree.task_tree.root.insert(self.session)
         result = self.session.scalars(select(PickUpWithContextView)
                                       .where(PickUpWithContextView.arm == "right")).all()
         self.assertEqual(result, [])
