+++ test/test_task_tree.py
@@ -1,9 +1,9 @@
-from pycram.designators.actions.actions import MoveTorsoActionPerformable, PickUpActionPerformable, \
+from pycram.designators.performables.actions import MoveTorsoActionPerformable, PickUpActionPerformable, \
     NavigateActionPerformable
 from pycram.datastructures.pose import Pose
 from pycram.process_module import simulated_robot
-import pycram.task
-from pycram.task import with_tree
+import pycram.tasktree
+from pycram.tasktree import with_tree
 import unittest
 import anytree
 from bullet_world_testcase import BulletWorldTestCase
@@ -26,21 +26,21 @@ class TaskTreeTestCase(BulletWorldTestCase):
 
     def setUp(self):
         super().setUp()
-        pycram.task.reset_tree()
+        pycram.tasktree.reset_tree()
 
     def test_tree_creation(self):
         """Test the creation and content of a task tree."""
         self.plan()
         # self.tearDownBulletWorld()
-        tt = pycram.task.task_tree
+        tt = pycram.tasktree.task_tree
 
         self.assertEqual(15, len(tt.root))
         self.assertEqual(10, len(tt.root.leaves))
 
         # check that all nodes succeeded
         for node in anytree.PreOrderIter(tt.root):
-            if not isinstance(node.action, pycram.task.NoOperation):
-                self.assertEqual(node.status, pycram.task.TaskStatus.SUCCEEDED)
+            if not isinstance(node.action, pycram.tasktree.NoOperation):
+                self.assertEqual(node.status, pycram.tasktree.TaskStatus.SUCCEEDED)
 
     def test_exception(self):
         """Test the tree with failing plans."""
@@ -49,38 +49,38 @@ class TaskTreeTestCase(BulletWorldTestCase):
         def failing_plan():
             raise pycram.plan_failures.PlanFailure("PlanFailure for UnitTesting")
 
-        pycram.task.reset_tree()
+        pycram.tasktree.reset_tree()
 
         self.assertRaises(pycram.plan_failures.PlanFailure, failing_plan)
 
-        tt = pycram.task.task_tree
+        tt = pycram.tasktree.task_tree
 
         for node in anytree.PreOrderIter(tt.root):
-            if not isinstance(node.action, pycram.task.NoOperation):
-                self.assertEqual(node.status, pycram.task.TaskStatus.FAILED)
+            if not isinstance(node.action, pycram.tasktree.NoOperation):
+                self.assertEqual(node.status, pycram.tasktree.TaskStatus.FAILED)
                 self.assertEqual(pycram.plan_failures.PlanFailure, type(node.reason))
 
     def test_execution(self):
         self.plan()
         self.world.reset_world()
-        tt = pycram.task.task_tree
+        tt = pycram.tasktree.task_tree
         # self.setUpBulletWorld(False)
         with simulated_robot:
             [node.action.perform() for node in tt.root.leaves]
 
     def test_simulated_tree(self):
-        with pycram.task.SimulatedTaskTree() as st:
+        with pycram.tasktree.SimulatedTaskTree() as st:
             self.plan()
-            tt = pycram.task.task_tree
+            tt = pycram.tasktree.task_tree
 
             self.assertEqual(15, len(tt.root))
             self.assertEqual(10, len(tt.root.leaves))
 
-        self.assertEqual(len(pycram.task.task_tree), 1)
+        self.assertEqual(len(pycram.tasktree.task_tree), 1)
 
     def test_to_sql(self):
         self.plan()
-        tt = pycram.task.task_tree
+        tt = pycram.tasktree.task_tree
         result = tt.root.to_sql()
         self.assertIsNotNone(result)
